+++ tests/test_issue11/xmake.lua
@@ -1,0 +1,29 @@
+---@diagnostic disable
+
+target("test", function()
+    add_rules("verilua")
+
+    on_config(function(target)
+        local sim = os.getenv("SIM") or "verilator"
+        if sim == "iverilog" then
+            target:set("toolchains", "@iverilog")
+        elseif sim == "vcs" then
+            target:set("toolchains", "@vcs")
+        elseif sim == "xcelium" then
+            target:set("toolchains", "@xcelium")
+        elseif sim == "verilator" then
+            target:set("toolchains", "@verilator")
+        else
+            raise("unknown simulator: %s", sim)
+        end
+    end)
+
+    add_files("top.v")
+
+    -- set_values("verilua.use_inertial_put", "1")
+
+    set_values("verilua.top", "top")
+    set_values("verilua.lua_main", "./main.lua")
+
+    set_values("verilator.flags", "--trace", "--no-trace-top")
+end)